changed file structures
