+++ App.js
@@ -2,7 +2,7 @@ import React from 'react';
 import {StyleSheet, View, Text, Image, Button} from 'react-native';
 import {NavigationContainer} from '@react-navigation/native';
 import {createNativeStackNavigator} from '@react-navigation/native-stack';
-import MainContainer from './navigation/MainContainer';
+import MainContainer from './src/Containers/MainContainer';
 
 function App() {
   return <MainContainer />;
