+++ src/Containers/MainContainer.js
@@ -5,14 +5,14 @@ import {createNativeStackNavigator} from '@react-navigation/native-stack';
 import {createBottomTabNavigator} from '@react-navigation/bottom-tabs';
 import Ionicons from 'react-native-vector-icons/Ionicons';
 
-//Screens
-import HomeScreen from './screens/HomeScreen';
-import FoodScreen from './screens/FoodScreen';
-import SleepScreen from './screens/SleepScreen';
-import BabyScreen from './screens/BabyScreen';
-import ListScreen from './screens/ListScreen';
+//Components
+import HomeScreen from '../Components/HomeScreen';
+import FoodScreen from '../Components/FoodScreen';
+import SleepScreen from '../Components/SleepScreen';
+import BabyScreen from '../Components/BabyScreen';
+import ListScreen from '../Components/ListScreen';
 import {HomeScreenNavigator, BabyScreenNavigator} from './StackNavigation';
-import SummaryScreen from './screens/SummaryScreen';
+import SummaryScreen from '../Components/SummaryScreen';
 
 //Screen names
 const homeName = "Home";
