+++ src/Services/FeedService.js
@@ -1,5 +1,5 @@
 
-const baseURL = 'http://192.168.0.55:8080/feeds'
+const baseURL = 'http://192.168.0.65:8080/feeds'
